added trust parameter in fun bank()

+++ app/src/main/java/com/thenewboston/kotlinsdk/network/models/BankConfigModel.kt
@@ -1,6 +1,5 @@
 package com.thenewboston.kotlinsdk.network.models
 
-
 import com.google.gson.annotations.SerializedName
 
 data class BankConfigModel(
@@ -8,7 +7,7 @@ data class BankConfigModel(
     @SerializedName("account_number")
     val accountNumber: String, // acc_no
     @SerializedName("default_transaction_fee")
-    val defaultTransactionFee: Int?, // 1
+    val defaultTransactionFee: Double?, // 1.0
     @SerializedName("ip_address")
     val ipAddress: String, // 143.110.137.54
     @SerializedName("node_identifier")
@@ -27,7 +26,7 @@ data class BankConfigModel(
         @SerializedName("daily_confirmation_rate")
         val dailyConfirmationRate: Any?, // null
         @SerializedName("default_transaction_fee")
-        val defaultTransactionFee: Int, // 1
+        val defaultTransactionFee: Double, // 1.0
         @SerializedName("ip_address")
         val ipAddress: String, // 54.183.17.224
         @SerializedName("node_identifier")
